Part 1
Here is part one. The graphs look pretty good except for the measles one. I think the measles one just has a transmission rate that is so high that basically everyone is getting infected which is why the graph looks so bad. Code for all three diseases should be exactly the same.