Added Part 3 for final project
Everything seems to work well. For step 3, the reason that using 10 days instead of 30 is better for approximating I(0) and beta is because this model approximates S to be constant and equal to S0. This approach is fine for the beginning of a pandemic, but it starts to break down when it progresses (more iterations). Code outputs a table showing true value, T = 10 approximation and T = 30 approximation for both I(0) and beta.